Added different lattice geometries
From Email to Pigeon:

- I renamed the lattice.{h,cpp} to baselattice.{h,cpp}. Thus the "Lattice" class
  is now a "BaseLattice", and is only there to provide functions that other
  classes (lattice geometries) inherit from. Some of its member functions are
  "pure virtual" and not defined, since they will be defined by the classes that
  inherit from this one, like the "moveAndBoundaryCheck()" and a new function
  called "distance()" which should return the distance "dr" from x,y,z coordinates.

- there's now a lattices.{h,cpp} where I currently define two classes: "Square"
  and "Honeycomb2d". Both of these inherit everything from the
  "BaseLattice"-class. But they define their own "moveAndBoundaryCheck()", and
  their own "distance()" function which computes the dr for that lattice.

You can look at how the "Square" lattice class is implemented in
lattices.{h,cpp}, since this is just the same old square lattice. I've made it
so that Square can be 1D,2D or 3D, depending on how it is constructed in
main().

The "Honeycomb2d" class is always 2D but with 6 directions to move in.

So in main() I now have a vector (called "lattices", just as before) with an
element for each thread, but now this is a vector of pointers to
BaseLattice-class, which will then at run time either be "Square" or
"Honeycomb2d" lattice, depending on the settings in the input file. This might
look strange to you, but since we don't know at compile time what kind of
lattice we want, it must be a vector of pointers to the base class which the
other classes inherit from. Thus it's now a std::vector<BaseLattice*> instead
of std::vector<Lattice>. But then I used the new "smart pointers" in C++11
standard called unique_ptr, which is a pointer that kills itself when it is no
longer needed, thus it is actually a std::vector<std::unique_ptr<BaseLattice>>.

Beyond that the change is not that big. As an example, I used to have
something like:
std::vector<Lattice> lattices.setJumpRate(...)
which is now:
std::vector<std::unique_ptr<BaseLattice>> lattices->setJumpRate(...)

Also, in the top of the main file there's an enum that enumerates the
different lattices that I can have, so I don't have to keep track on that "4
means honeycomb-lattice".

So, it seems to run now, but when I run the honeycomb lattice ("4" in input
file), the displacement seems to not start from 0, but some large value.

So I used the formula you gave me (the way I understood it), as you can see in
lattices.cpp where I have the function "distance()" which does:

$d = \sqrt{ ([x-x_0] + \frac{1}{2}[y-y_0])^2 + \frac{3}{4}(y-y_0)^2 }$

where x,y are the current coordinates and x_0,y_0 the coordinates at t=0. I
assume something is wrong here since the displacement should start at 0, if
you do a simulation, you'll see what I mean, by doing something like:

make
./prog -r input.ini -w out.dat

I'll think about this bug some more. Seems to be related to the size of the
lattice, but not the number of "ensembles" / runs.