remove Direction and Jump classes, use vector instead in Particle class
Have a vector in the Particle class instead where each element is the jumprate
in a certain direction. This makes for less convoluted code and will allow for
arbitrarily many directions, de-coupled from the dimension of the lattice.

(Consider honeycomb which has 6 direction on a 2D lattice.)

Also remove the BaseLattice::jumpRate_ vector since this data is stored in the
Particle vector.